Initial codex UI
Added to its own scene (and the "latest_demo" scene) a UI meant to track the Codex feature. Currently, it will enable if the player pressed the 'C' key, which currently is the key in the input map to toggle the codex. Will likely change the layout of the codex UI based of off design changes!