Refactored sword screen shake
The sword screen shake was overly coupled with the sword; the code could not be re-used for any other screen shake.
It was also confusing for the sword to be changing the camera like that. The bulk of the code has been moved to the a script attached to the camera (see world/camera/main_cam).