Added Two Buttons
Changed the layout of the codex menu (which when playing the main scene you can access by pressing 'c' to be a little more like a book, and added two buttons. When the button is pressed, it prints the "index" of the button, which is 1 for the top one and 2 for the bottom one. Currently, to add more buttons, you simply need to add one button node to the "CodexEnemyControl" node, add its unique name to the @onready section of variables, and call in the InitButtons() function in the codex.gd script .SetIndex(index) on the button, which should make it so if you press the new button it'll print out the index you assigned to it in the console.

Will likely need a better way to organize the buttons, as having to add a new variable and have it in the script for EVERY potential codex entry will be exhausting, for me and for the line length of the file. Ideally by next meeting that will be sorted out, and I can actually begin to look at making the different pages and loading the codex entries.