Improved game over screen and started working on title screen
I added an Image of a sleeping dog to the game over screen. I also changed the file change to the UID instead of its root file just in case any of the files get moved around. Lastly, I started to work on the main title screen of the game (basic and simple for now).