Pressure Plate Compatibility
Targets can now use a pressure plate to activate. The pressure plate takes precedent over the area2d

If the target has an area2d only, if the player steps into the area2d, that target will run

If the target has a pressure plate only, if the player steps on the pressure plate, that target will run

If the target has a pressure plate and an area2d, the pressure plate takes precedent and the pressure plate will run the targets

If a target A has only the area2d and target B has both area2d and pressure plate. If the player is in the area2d, target A will run, but A and B will play the flash animation

If the target has neither area2d or pressure plate, then the target will simply not run nor will it flash

Pressure plate takes precedents, so if Target A is area2d and Target B is pressure plate AND area2d if Player is on pressure plate, Target B will run and flash. If player is in area 2d, Target A will run and A and B will flash

There are sure to be some more combinations, but just keep in mind for now that the pressure plate takes precedents and the flash is determinant on that