Target Groups
Added the ability to group which targets should play the animation instead of all of the targets. The movement is still dependent on the Area2d and Pressure plate, with pressure plate taking Precedent over Area2d, they can both be unassigned in the inspector and both assigned or just one assigned. But the main thing that needs to be kept in mind is that the pressure plate takes precedent, so if you have a pressure plate and an area2d on one target and you're expecting the area2d to affect it, it wont.