Added Sword Detection Visual / Removed From Hittable Group
Created an area2d around the target to detect if the sword entered the area to detect a hit. At this current time, having it in the hittable group seemed unreliable. Perhaps later down the road it should change, but for now, it shall be an area2d  detection

The sword will flash for ~.5 seconds, indicating it has been hit. This is the same animation for when the timer is about to run out. Perhaps this will change to get a more noticeable hit with the sword

Ultimately the Target is pretty much done