Big Codex Update!
Added the ability for the codex to automatically adapt to certain entries being locked between other entries (for example - if entry 1 is unlocked, 2 is locked, and 3 is unlocked, it will set the left page to be entry 1 and the right to be entry 2). There is one final thing to fix - the page number on the bottom corner of each page corresponds to the index of the entry in the whole dictionary, not the unlocked ones. If we use the previous example of 1 being unlocked, 2 being locked, and 3 being unlocked, the page number at the bottom corner of 3's page will be 3, when it should be 2 since it is the 2nd unlocked entry of the ones in the dictionary.