Final Target Practice
Target Completion:

The target is able to move along a 2d path.
Who ever is placing the targets, they are able to attach an area2d with collision and or the pressure plate.
They can also adjust the speed of the target, how long the target runs for, when the flash indicator starts and also the scale up speed for how fast the flash animation plays.

There is an option to change the sprite within any scene (Not tested). If there is a target sprite, best to change it within the target scene itself

There are also groups you can assign the targets to. It can be any string, as long as the group names are consistent upon the targets you want to connect. This mainly connects the flash animation with "active" targets. If you want the groups to move at the same time, then you should connect either the same area2d or pressure plate