Fixed the bomb-sword issue
Fixed the issue of the bomb exploding as soon as the sword hit it, regardless of the speed of the bomb. This allowed the bomb to potentially explode as soon as the player threw it, if either the player also threw their sword at the same time, or the sword was coming back to the player. The fix involves setting the velocity of the bomb as just zero if its velocity is small enough, which does change the distance the bomb travels (by roughly 3-5% or so, to change it just modify the two numbers in the less than inequality checks in the _physics_process() function) but also ensures the bomb will eventually stop as we simply multiply its speed by 90% every frame, allowing the sword to explode it once it stops.