Side Quest for The Librarian
This is a brief description of what I (Peter Dang) had prepared for this side quest.

This side quest starts with interacting with the Librarian/and his wife at the Library. Once the player accepts the quest, they must "fetch" 3 books via interaction by dialogue with a few villagers: The Alchemist, The Doctor, and The Gossipy Elder. Once the player has selected to collect a book, a book counter could be implemented to increment after each specific dialogue from each villager. Once the player has 3 books, they should traverse back to the Librarian and receive appropriate XP and currency.